Update Fortnite names decryption | Add defs.h

+++ Dumper/memory.h
@@ -1,11 +1,10 @@
 #pragma once
 #include <windows.h>
-#include <cstdint>
+#include "defs.h"
 
-extern uint64_t Base;
+extern uint64 Base;
 
-bool Read(void* address, void* buffer, size_t size);
-
+bool Read(void* address, void* buffer, uint64 size);
 template<typename T>
 T Read(void* address) 
 {
@@ -13,5 +12,4 @@ T Read(void* address)
 	Read(address, &buffer, sizeof(T));
 	return buffer;
 }
-
-bool ReaderInit(uint32_t pid);
+bool ReaderInit(uint32 pid);
